writeExternalNodes: adding bandwidth and delay
Currently --bandwidth from graphgen will modify clickGen,
but it fails to modify the ns file from nsGen.  A more substantial
fix should be a means to specify links, whether interactive or via
file, but for now, this is a change to modify all links for the ns
file.  I did not verify how click does it - but this is the only
option.

This is a `quick fix`, there are more pythonic approaches, and
this does violate DRY, so another fix may be required.

Lincoln