nsGen.py: re-write code for logic, performance, lint
This commit should, be broken into multiple commits as it changes
the entire nsGen.py file.  These changes make the code a bit longer,
but make the code faster, reduce the amount of duplicate code,
and linted based on the pylintrc file in the travis-ci branch.

There should be a review of the file inself to continue reducing the
amount of duplicate code by adding more boilerplate.

This code was tested and verified to generate working ns topology
files that can be run on deter.

Lincoln