graphGen: add initial travisCI configuration file
Added a basic example that will verify the generated files dont
differ.  If there is a difference either the commit broke the
functionality, or the master file (either ns or out) must be
updated to reflect the changes.

This is the first basic implementation for quality.  More tests,
as well as using pytest or unittest should be used when able.

Lincoln