adding tcl syntax checker
Using nagelfar tcl syntax checker as another tool to verify the
correctness of the generated tcl (ns) file for deterlab.

Add a simple fix to ns_gen file to correctly format/output the
correct file.

Next step is: continued code cleanup.  Beyond that, need to add
testing for click module, and separate the code for unit-testing.